refactor vulnerability ingester code
