+++ lunatrace/bsl/license-worker/cmd/lpt/vulnerability/vulnerability.go
@@ -12,85 +12,44 @@
 package vulnerability
 
 import (
-  "context"
-  "github.com/lunasec-io/lunasec/lunatrace/bsl/license-worker/pkg/vulnerability"
-  "github.com/rs/zerolog/log"
-  "github.com/urfave/cli/v2"
-  "go.uber.org/fx"
-  "io/fs"
-  "os"
-  "path/filepath"
+	"context"
+	"github.com/lunasec-io/lunasec/lunatrace/bsl/license-worker/pkg/vulnerability"
+	"github.com/urfave/cli/v2"
+	"go.uber.org/fx"
 
-  "github.com/lunasec-io/lunasec/lunatrace/bsl/license-worker/pkg/clifx"
+	"github.com/lunasec-io/lunasec/lunatrace/bsl/license-worker/pkg/clifx"
 )
 
 type Params struct {
-  fx.In
+	fx.In
 
-  Loader vulnerability.Loader
+	Ingester vulnerability.FileAdvisoryIngester
 }
 
 func NewCommand(p Params) clifx.CommandResult {
-  return clifx.CommandResult{
-    Command: &cli.Command{
-      Name: "vulnerability",
-      Subcommands: []*cli.Command{
-        {
-          Name:  "ingest",
-          Usage: "[file or directory]",
-          Flags: []cli.Flag{
-            &cli.StringFlag{
-              Name:     "source",
-              Usage:    "Where the vulnerabilities have been sourced from.",
-              Required: true,
-            },
-          },
-          Action: func(ctx *cli.Context) error {
-            fileLocation := ctx.Args().First()
+	return clifx.CommandResult{
+		Command: &cli.Command{
+			Name: "vulnerability",
+			Subcommands: []*cli.Command{
+				{
+					Name:  "ingest",
+					Usage: "[file or directory]",
+					Flags: []cli.Flag{
+						&cli.StringFlag{
+							Name:     "source",
+							Usage:    "Where the vulnerabilities have been sourced from.",
+							Required: true,
+						},
+					},
+					Action: func(ctx *cli.Context) error {
+						advisoryLocation := ctx.Args().First()
+						source := ctx.String("source")
 
-            fileInfo, err := os.Stat(fileLocation)
-            if err != nil {
-              log.Error().
-                Err(err).
-                Msg("unable to get file info for file location")
-              return err
-            }
-
-            var files []string
-
-            if fileInfo.IsDir() {
-              err = filepath.WalkDir(fileLocation, func(path string, d fs.DirEntry, err error) error {
-                if d.IsDir() {
-                  return nil
-                }
-                files = append(files, path)
-                return nil
-              })
-              if err != nil {
-                log.Error().
-                  Err(err).
-                  Msg("unable to load files from directory file location")
-                return err
-              }
-            } else {
-              files = append(files, fileLocation)
-            }
-
-            loadCtx := context.Background()
-            for _, filename := range files {
-              err = p.Loader.Load(loadCtx, filename)
-              if err != nil {
-                log.Warn().
-                  Err(err).
-                  Str("filename", filename).
-                  Msg("failed to load vulnerability")
-                continue
-              }
-            }
-            return nil
-          },
-        },
-      },
-    },
-  }
+						ingestCtx := context.Background()
+						return p.Ingester.Ingest(ingestCtx, source, advisoryLocation)
+					},
+				},
+			},
+		},
+	}
 }
