--- lunatrace/bsl/license-worker/pkg/vulnerability/mapper.go
@@ -1,466 +1,0 @@
-// Copyright by LunaSec (owned by Refinery Labs, Inc)
-//
-// Licensed under the Business Source License v1.1 
-// (the "License"); you may not use this file except in compliance with the
-// License. You may obtain a copy of the License at
-//
-// https://github.com/lunasec-io/lunasec/blob/master/licenses/BSL-LunaTrace.txt
-//
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-package vulnerability
-
-import (
-  "encoding/json"
-  "github.com/lunasec-io/lunasec/lunatrace/bsl/license-worker/pkg/vulnerability/schema"
-  "github.com/lunasec-io/lunasec/lunatrace/cli/gql"
-  "github.com/lunasec-io/lunasec/lunatrace/cli/gql/types"
-  "github.com/lunasec-io/lunasec/lunatrace/cli/pkg/util"
-  "github.com/rs/zerolog/log"
-  "golang.org/x/exp/slices"
-  "strings"
-  "time"
-)
-
-// MapOsvToGraphql converts an OSV formatted vulnerability to the Graphql representation of a vulnerability.
-// Example OSV: https://github.com/github/advisory-database/blob/main/advisories/github-reviewed/2021/12/GHSA-jfh8-c2jp-5v3q/GHSA-jfh8-c2jp-5v3q.json
-func MapOsvToGraphql(source string, v *schema.OsvSchema) (graphqlInsert *gql.Vulnerability_insert_input, err error) {
-  osvSchema, err := serializeToJsonRawMessage(v)
-  if err != nil {
-    log.Error().
-      Err(err).
-      Str("id", v.Id).
-      Msg("unable to serialize osv schema")
-    return
-  }
-
-  databaseSpecific, err := serializeToJsonRawMessage(v.DatabaseSpecific)
-  if err != nil {
-    log.Error().
-      Err(err).
-      Str("id", v.Id).
-      Msg("unable to serialize database specific info")
-    return
-  }
-
-  modifiedTime, err := time.Parse(time.RFC3339, v.Modified)
-  if err != nil {
-    log.Error().
-      Err(err).
-      Str("id", v.Id).
-      Msg("unable to parse modified time")
-    return
-  }
-
-  mergedAffected := mergeOsvPackages(v.Affected)
-
-  graphqlInsert = &gql.Vulnerability_insert_input{
-    Modified:          util.Ptr(modifiedTime),
-    Published:         parseOptionalTime(v.Published),
-    Withdrawn:         parseOptionalTime(v.Withdrawn),
-    Affected:          mapAffected(mergedAffected),
-    Credits:           mapCredits(v.Credits),
-    Database_specific: &databaseSpecific,
-    Details:           v.Details,
-    Equivalents:       mapEquivalents(v.Aliases),
-    References:        mapReferences(v.References),
-    Severities:        mapSeverities(source, v.Severity),
-    Source:            util.Ptr(source),
-    Source_id:         util.Ptr(v.Id),
-    Summary:           v.Summary,
-    Upstream_data:     util.Ptr(osvSchema),
-  }
-  return
-}
-
-func stringSlicePostgresqlArray(slice []string) string {
-  return "{" + strings.Join(slice, ",") + "}"
-}
-
-func serializeToJsonRawMessage(i interface{}) (json.RawMessage, error) {
-  marshalled, err := json.Marshal(i)
-  if err != nil {
-    return nil, err
-  }
-  return marshalled, nil
-}
-
-// OSV affected packages are split across multiple entries in the affected array. These packages can be merged
-// to make querying for affected versions more straight forward.
-func mergeOsvPackages(affected []schema.OsvSchemaAffectedElem) []schema.OsvSchemaAffectedElem {
-  affectedLookup := map[string]schema.OsvSchemaAffectedElem{}
-
-  for _, affectedElem := range affected {
-    affectedPackage := affectedElem.Package
-
-    // TODO (cthompson) what do we want to do if no package is provided? should we just be ignoring it?
-    if affectedPackage == nil {
-      continue
-    }
-
-    lookupKey := affectedPackage.Ecosystem + affectedPackage.Name
-    if existingElem, ok := affectedLookup[lookupKey]; ok {
-      existingElem.Versions = append(existingElem.Versions, affectedElem.Versions...)
-      existingElem.Ranges = append(existingElem.Ranges, affectedElem.Ranges...)
-      affectedLookup[lookupKey] = existingElem
-      continue
-    }
-    affectedLookup[lookupKey] = affectedElem
-  }
-
-  newAffected := make([]schema.OsvSchemaAffectedElem, 0, len(affectedLookup))
-  for _, affectedElem := range affectedLookup {
-    newAffected = append(newAffected, affectedElem)
-  }
-  return newAffected
-}
-
-func mapReferences(references []schema.OsvSchemaReferencesElem) *gql.Vulnerability_reference_arr_rel_insert_input {
-  if len(references) == 0 {
-    return nil
-  }
-
-  data := make([]*gql.Vulnerability_reference_insert_input, 0, len(references))
-  for _, element := range references {
-    data = append(data, mapReferenceElement(element))
-  }
-  return &gql.Vulnerability_reference_arr_rel_insert_input{
-    Data: data,
-    On_conflict: &gql.Vulnerability_reference_on_conflict{
-      Constraint: gql.Vulnerability_reference_constraintReferenceVulnerabilityIdTypeUrlIdx,
-      Update_columns: []gql.Vulnerability_reference_update_column{
-        gql.Vulnerability_reference_update_columnVulnerabilityId,
-        gql.Vulnerability_reference_update_columnType,
-        gql.Vulnerability_reference_update_columnUrl,
-      },
-    },
-  }
-}
-
-func mapReferenceElement(reference schema.OsvSchemaReferencesElem) *gql.Vulnerability_reference_insert_input {
-  return &gql.Vulnerability_reference_insert_input{
-    Type: util.Ptr(mapReferenceType(reference.Type)),
-    Url:  util.Ptr(reference.Url),
-  }
-}
-
-func mapSeverities(source string, severities []schema.OsvSchemaSeverityElem) *gql.Vulnerability_severity_arr_rel_insert_input {
-  if len(severities) == 0 {
-    return nil
-  }
-
-  data := make([]*gql.Vulnerability_severity_insert_input, 0, len(severities))
-  for _, element := range severities {
-    data = append(data, mapSeverityElement(source, element))
-  }
-  return &gql.Vulnerability_severity_arr_rel_insert_input{
-    Data: data,
-    On_conflict: &gql.Vulnerability_severity_on_conflict{
-      Constraint: gql.Vulnerability_severity_constraintSeverityVulnerabilityIdSourceTypeIdx,
-      Update_columns: []gql.Vulnerability_severity_update_column{
-        gql.Vulnerability_severity_update_columnScore,
-      },
-    },
-  }
-}
-
-func mapSeverityElement(source string, severity schema.OsvSchemaSeverityElem) *gql.Vulnerability_severity_insert_input {
-  return &gql.Vulnerability_severity_insert_input{
-    Score:  util.Ptr(severity.Score),
-    Source: util.Ptr(source),
-
-    // TODO (cthompson) it looks like for severity only CVSS_V3 is defined https://github.com/ossf/osv-schema/blob/main/validation/schema.json#L43
-    // do we want to make the type column an enum?
-    Type: util.Ptr(string(severity.Type)),
-  }
-}
-
-func mapEquivalents(aliases []string) *gql.Vulnerability_equivalent_arr_rel_insert_input {
-  if len(aliases) == 0 {
-    return nil
-  }
-
-  data := make([]*gql.Vulnerability_equivalent_insert_input, 0, len(aliases))
-  for _, element := range aliases {
-    data = append(data, mapEquivalentElement(element))
-  }
-  return &gql.Vulnerability_equivalent_arr_rel_insert_input{
-    Data: data,
-    On_conflict: &gql.Vulnerability_equivalent_on_conflict{
-      Constraint: gql.Vulnerability_equivalent_constraintEquivalentABIdx,
-      Update_columns: []gql.Vulnerability_equivalent_update_column{
-        gql.Vulnerability_equivalent_update_columnA,
-        gql.Vulnerability_equivalent_update_columnB,
-      },
-    },
-  }
-}
-
-func mapEquivalentElement(alias string) *gql.Vulnerability_equivalent_insert_input {
-  source := determineSourceFromId(alias)
-  return &gql.Vulnerability_equivalent_insert_input{
-    Vulnerability: nil,
-    VulnerabilityByB: &gql.Vulnerability_obj_rel_insert_input{
-      Data: &gql.Vulnerability_insert_input{
-        Source:    util.Ptr(source),
-        Source_id: util.Ptr(alias),
-      },
-      On_conflict: &gql.Vulnerability_on_conflict{
-        Constraint: gql.Vulnerability_constraintVulnerabilitySourceSourceIdIdx,
-        Update_columns: []gql.Vulnerability_update_column{
-          gql.Vulnerability_update_columnSource,
-          gql.Vulnerability_update_columnSourceId,
-        },
-      },
-    },
-  }
-}
-
-func determineSourceFromId(id string) string {
-  if strings.HasPrefix(id, "CVE-") {
-    return "nvd"
-  }
-  if strings.HasPrefix(id, "GHSA-") {
-    return "github"
-  }
-  panic("unable to determine source from id: " + id)
-}
-
-func mapCredits(credits []schema.OsvSchemaCreditsElem) *gql.Vulnerability_credit_arr_rel_insert_input {
-  if len(credits) == 0 {
-    return nil
-  }
-
-  data := make([]*gql.Vulnerability_credit_insert_input, 0, len(credits))
-  for _, element := range credits {
-    data = append(data, mapCreditElement(element))
-  }
-  return &gql.Vulnerability_credit_arr_rel_insert_input{
-    Data: data,
-    On_conflict: &gql.Vulnerability_credit_on_conflict{
-      Constraint: gql.Vulnerability_credit_constraintCreditVulnerabilityIdName,
-      Update_columns: []gql.Vulnerability_credit_update_column{
-        gql.Vulnerability_credit_update_columnVulnerabilityId,
-        gql.Vulnerability_credit_update_columnName,
-      },
-    },
-  }
-}
-
-func mapCreditElement(credit schema.OsvSchemaCreditsElem) *gql.Vulnerability_credit_insert_input {
-  formattedContact := stringSlicePostgresqlArray(credit.Contact)
-  return &gql.Vulnerability_credit_insert_input{
-    Contact: util.Ptr(formattedContact),
-    Name:    util.Ptr(credit.Name),
-  }
-}
-
-func parseOptionalTime(optionalTime *string) *time.Time {
-  if optionalTime == nil {
-    return nil
-  }
-
-  parsedTime, err := time.Parse(time.RFC3339, *optionalTime)
-  if err != nil {
-    panic(err)
-  }
-  return &parsedTime
-}
-
-func mapAffected(affected []schema.OsvSchemaAffectedElem) *gql.Vulnerability_affected_arr_rel_insert_input {
-  if len(affected) == 0 {
-    return nil
-  }
-
-  data := make([]*gql.Vulnerability_affected_insert_input, 0, len(affected))
-  for _, element := range affected {
-    data = append(data, mapAffectedElement(element))
-  }
-  return &gql.Vulnerability_affected_arr_rel_insert_input{
-    Data: data,
-    On_conflict: &gql.Vulnerability_affected_on_conflict{
-      Constraint: gql.Vulnerability_affected_constraintAffectedVulnerabilityIdPackageIdIdx,
-      Update_columns: []gql.Vulnerability_affected_update_column{
-        gql.Vulnerability_affected_update_columnDatabaseSpecific,
-        gql.Vulnerability_affected_update_columnEcosystemSpecific,
-        gql.Vulnerability_affected_update_columnPackageId,
-        gql.Vulnerability_affected_update_columnVulnerabilityId,
-      },
-    },
-  }
-}
-
-func mapAffectedElement(element schema.OsvSchemaAffectedElem) *gql.Vulnerability_affected_insert_input {
-  databaseSpecific, err := serializeToJsonRawMessage(element.DatabaseSpecific)
-  if err != nil {
-    panic(err)
-  }
-
-  ecosystemSpecific, err := serializeToJsonRawMessage(element.EcosystemSpecific)
-  if err != nil {
-    panic(err)
-  }
-
-  // TODO (cthompson) update package schema to accept repo
-  rangeEvents, _ := mapAffectedRanges(element.Ranges)
-
-  if element.Package == nil {
-    panic("package is nil")
-  }
-
-  affectedPackage := &gql.Package_obj_rel_insert_input{
-    Data: &gql.Package_insert_input{
-      Custom_registry: util.Ptr(""),
-      Name:            util.Ptr(element.Package.Name),
-      Package_manager: util.Ptr(mapEcosystemToPackageManager(element.Package.Ecosystem)),
-    },
-    On_conflict: &gql.Package_on_conflict{
-      Constraint: gql.Package_constraintPackagePackageManagerCustomRegistryNameIdx,
-      Update_columns: []gql.Package_update_column{
-        gql.Package_update_columnName,
-        gql.Package_update_columnPackageManager,
-        gql.Package_update_columnCustomRegistry,
-      },
-    },
-  }
-
-  return &gql.Vulnerability_affected_insert_input{
-    Affected_range_events: rangeEvents,
-    Affected_versions:     mapAffectedVersion(element.Versions),
-    Database_specific:     util.Ptr(databaseSpecific),
-    Ecosystem_specific:    util.Ptr(ecosystemSpecific),
-    Package:               affectedPackage,
-  }
-}
-
-func mapAffectedRanges(ranges []schema.OsvSchemaAffectedElemRangesElem) (*gql.Vulnerability_affected_range_event_arr_rel_insert_input, string) {
-  var repo string
-
-  if len(ranges) == 0 {
-    return nil, repo
-  }
-
-  // A lookup to determine if an event has already been added to the insertion.
-  // This will prevent duplicate event inserts.
-  eventLookup := map[string]schema.OsvSchemaAffectedElemRangesElemEventsElem{}
-
-  data := make([]*gql.Vulnerability_affected_range_event_insert_input, 0, len(ranges))
-  for _, rangeElement := range ranges {
-    // TODO (cthompson) how should we handle multiple defined repos?
-    if rangeElement.Repo != nil {
-      repo = *rangeElement.Repo
-    }
-    data = append(data, mapAffectedRangeElement(eventLookup, rangeElement)...)
-  }
-
-  affectedRangeInput := &gql.Vulnerability_affected_range_event_arr_rel_insert_input{
-    Data: data,
-    On_conflict: &gql.Vulnerability_affected_range_event_on_conflict{
-      Constraint: gql.Vulnerability_affected_range_event_constraintAffectedRangeEventAffectedIdTypeEventVersionIdx,
-      Update_columns: []gql.Vulnerability_affected_range_event_update_column{
-        gql.Vulnerability_affected_range_event_update_columnAffectedId,
-        gql.Vulnerability_affected_range_event_update_columnType,
-        gql.Vulnerability_affected_range_event_update_columnEvent,
-        gql.Vulnerability_affected_range_event_update_columnVersion,
-      },
-    },
-  }
-  return affectedRangeInput, repo
-}
-
-func mapAffectedRangeElement(
-  eventLookup map[string]schema.OsvSchemaAffectedElemRangesElemEventsElem,
-  element schema.OsvSchemaAffectedElemRangesElem,
-) []*gql.Vulnerability_affected_range_event_insert_input {
-  databaseSpecific, err := serializeToJsonRawMessage(element.DatabaseSpecific)
-  if err != nil {
-    panic(err)
-  }
-
-  rangeType := mapRangeEventTypeToAffectedRangeType(element.Type)
-
-  var events []*gql.Vulnerability_affected_range_event_insert_input
-  for _, event := range element.Events {
-    for name, value := range event {
-      eventValue := getEventValue(value)
-      eventKey := name + eventValue
-
-      if _, ok := eventLookup[eventKey]; ok {
-        // this absolutely matches an existing event, skip adding this event again.
-        continue
-      }
-
-      rangeEvent := &gql.Vulnerability_affected_range_event_insert_input{
-        Event:             util.Ptr(name),
-        Type:              util.Ptr(rangeType),
-        Version:           util.Ptr(eventValue),
-        Database_specific: util.Ptr(databaseSpecific),
-      }
-      events = append(events, rangeEvent)
-    }
-  }
-  return events
-}
-
-func getEventValue(value interface{}) string {
-  switch t := value.(type) {
-  case string:
-    return t
-  default:
-    panic("unknown type for event value")
-  }
-}
-
-func mapAffectedVersion(versions []string) *gql.Vulnerability_affected_version_arr_rel_insert_input {
-  if len(versions) == 0 {
-    return nil
-  }
-
-  data := make([]*gql.Vulnerability_affected_version_insert_input, 0, len(versions))
-  for _, version := range versions {
-    data = append(data, &gql.Vulnerability_affected_version_insert_input{
-      Version: util.Ptr(version),
-    })
-  }
-  return &gql.Vulnerability_affected_version_arr_rel_insert_input{
-    Data: data,
-    On_conflict: &gql.Vulnerability_affected_version_on_conflict{
-      Constraint: gql.Vulnerability_affected_version_constraintAffectedVersionAffectedIdTypeVersion,
-      Update_columns: []gql.Vulnerability_affected_version_update_column{
-        gql.Vulnerability_affected_version_update_columnAffectedId,
-        gql.Vulnerability_affected_version_update_columnVersion,
-      },
-    },
-  }
-}
-
-func mapReferenceType(osvReferenceType schema.OsvSchemaReferencesElemType) types.ReferenceType {
-  normalizedOsvReferenceType := strings.ToLower(string(osvReferenceType))
-  referenceType := types.ReferenceType(normalizedOsvReferenceType)
-  if !slices.Contains(types.ReferenceTypes, referenceType) {
-    panic("unable to determine reference type for " + normalizedOsvReferenceType)
-  }
-  return referenceType
-}
-
-// TODO (cthompson) need to be able to map this string to a package manager
-func mapEcosystemToPackageManager(ecosystem string) types.PackageManager {
-  normalizedEcosystem := strings.ToLower(ecosystem)
-  normalizedEcosystem = strings.ReplaceAll(normalizedEcosystem, ".", "_")
-
-  packageManager := types.PackageManager(normalizedEcosystem)
-  if !slices.Contains(types.PackageManagers, packageManager) {
-    panic("unable to find package manager for " + normalizedEcosystem)
-  }
-  return packageManager
-}
-
-func mapRangeEventTypeToAffectedRangeType(eventType schema.OsvSchemaAffectedElemRangesElemType) types.AffectedRangeType {
-  normalizedEventType := strings.ToLower(string(eventType))
-  rangeType := types.AffectedRangeType(normalizedEventType)
-  if !slices.Contains(types.AffectedRangeTypes, rangeType) {
-    panic("unable to find range type for " + normalizedEventType)
-  }
-  return rangeType
-}
