+++ lunatrace/cli/gql/upsertVulnerabilities.graphql
@@ -1,0 +1,7 @@
+mutation UpsertVulnerabilities($vulnerabilities: [vulnerability_insert_input!]!, $vulnerability_on_conflict: vulnerability_on_conflict!) {
+  insert_vulnerability(objects: $vulnerabilities, on_conflict: $vulnerability_on_conflict) {
+    returning {
+      id
+    }
+  }
+}
