--- lunatrace/cli/gql/upsertVulnerability.graphql
@@ -1,5 +1,0 @@
-mutation UpsertVulnerability($vulnerability: vulnerability_insert_input!, $on_conflict: vulnerability_on_conflict!) {
-  insert_vulnerability_one(object: $vulnerability, on_conflict: $on_conflict) {
-    id
-  }
-}
